+++ tests/Rust/tests/src/StringTests.fs
@@ -135,14 +135,22 @@ let ``StringBuilder.AppendFormat with provider works`` () =
 let ``StringBuilder.Chars works`` () =
     let sb = System.Text.StringBuilder()
                         .Append("abc")
+                        .Append("def")
+    sb.Chars(0) |> equal 'a'
     sb.Chars(1) |> equal 'b'
+    sb.Chars(2) |> equal 'c'
+    sb.Chars(3) |> equal 'd'
+    sb.Chars(4) |> equal 'e'
+    sb.Chars(5) |> equal 'f'
+
 
 [<Fact>]
 let ``StringBuilder.Chars throws when index is out of bounds`` () =
+    let sb = System.Text.StringBuilder()
+                        .Append("abc")
     throwsAnyError <| fun () ->
-        let sb = System.Text.StringBuilder()
-                            .Append("abc")
         sb.Chars(-1) |> ignore
+    throwsAnyError <| fun () ->
         sb.Chars(3) |> ignore
 
 [<Fact>]
@@ -166,6 +174,10 @@ let ``StringBuilder index setter works`` () =
                         .Append("abc")
     sb[1] <- 'x'
     sb.ToString() |> equal "axc"
+    throwsAnyError <| fun () ->
+        sb[-1] <- 'y'
+    throwsAnyError <| fun () ->
+        sb[3] <- 'z'
 
 [<Fact>]
 let ``kprintf works`` () =